pyproject.toml
[tool.poetry]
name = "ai-master"
version = "0.1.0"
description = ""
authors = ["AhernShaia <gemgt4gtii@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
langchain-core = "^0.1.50"
langchain-openai = "^0.1.6"
langchain-community = "^0.0.36"
langchain = "^0.1.17"
redis = "^5.0.4"
qdrant-client = "^1.9.1"
uvicorn = "^0.29.0"
numexpr = "^2.10.0"
langchainhub = "^0.1.15"
google-search-results = "^2.4.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
